+++ src/core/Screen.h
@@ -6,15 +6,15 @@
 
 class Screen {
 private:
-    double width, height;
+    int width, height;
     bool windowed;
     std::unique_ptr<GLFWwindow, decltype(&glfwDestroyWindow)> pwindow;
 
 public:
-    Screen(double w, double h, bool windowed);
+    Screen(int w, int h, bool windowed);
 
     float aspect() const {
-        return width / height;
+        return static_cast<float>(width) / static_cast<float>(height);
     }
 
     GLFWwindow* getWin() const {
